Ensure that `BRANCHSTART` commands are present on all branches
Before, we assumed that #BRANCHSTART would always get applied to the #N (normal) branch. This meant we could cache the #BRANCHSTART event in `converters.py`, then apply it to all subsequent branches.

But, following the previous refactor, #BRANCHSTART will now get immediately applied to the current branch, which will likely be #M, the last branch in the (#N, #E, #M) grouping. This means that we can no longer use the "cache the first event" strategy of processing the #BRANCHSTART command, since #M comes last, and thus #N and #E won't have their #BRANCHSTART condition applied to them.

To fix this, we simply set currentBranch to 'all' for #BRANCHSTART. This is safe to do, since currentBranch will get reset to 'normal' the moment an #N command is encountered, anyway.

This has the added benefit of removing the need to cache during `converters.py`, so we remove that logic, too.