`parsers.py`: Remove `targetBranch` logic
This logic came from the TJA parsing code in parseTJA.js, and has been preserved for the duration of the project. From what I can understand, it is meant to track which branch should be used after a #BRANCHEND command is encountered. Supposedly, this varies based on how the #BRANCHSTART was written.

However, if a #BRANCHEND is encountered, shouldn't the remaining data be appended to _all_ branches, and not just the "target branch"? Otherwise, the branches will not have the same number of measures, which is disallowed in fumens.

If my understanding is correct, then #BRANCHEND should reset the current branch to 'all'. And, once we make this change, all of the remaining `targetBranch` logic can be removed as well.