`parsers.py`: Refactor `current_branch == 'all'` logic
Fumens require that all branches have the same length, even when branching isn't
relevant (i.e. at the start of the song before the first BRANCHSTART, or after a
BRANCHEND). So, in the past, I simply duplicated the note data across all branches.

However, this causes a bug when balloon notes are duplicated; we get extra reads of
the BALLOON: field, and the values get mis-attributed. So, we need to be able to
modify this "dummy" note data before it gets duplicated. To do that, we need to rework
the 'current_branch == 'all'` logic to separate out the note data writes for the
other branches, so that we can modify them in-place.